stash apply: report status correctly even in a worktree's subdirectory
When Git wants to spawn a child Git process inside a worktree's
subdirectory, we need to take care of specifying the work tree's
top-level directory explicitly because it cannot be discovered: the
current directory is _not_ the top-level directory of the work tree, and
neither is it inside the parent directory of `GIT_DIR`.

This fixes the problem where `git stash apply` would report pretty much
everything deleted or untracked when run inside a worktree's
subdirectory.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>