Add a helper to reverse index_pos_to_insert_pos()
We have just introduced the helper `index_pos_to_insert_pos()` to help
avoiding underflows when returning `-1 - pos` for cases where we want to
return an insert position, using the ones' complement (as `int`).

As pointed out during the review of the patch series that introduced it,
this helper wants to be accompanied by a helper that reverse that ones'
complement, for clarity. This patch does exactly that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>