documentation: remove empty doc files
Remove empty and redundant documentation files from the
Documentation/technical/ directory.

The empty doc files included only TODO messages with no documentation for
years. Instead an approach is being taken to keep all doc beside the code
in the relevant header files.
Having empty doc files is confusing and disappointing to anybody looking
for information, besides having the documentation in header files makes it
easier for developers to find the information they are looking for.

here's a list of the files removed and if the info can be found in the
corresponding header file:
1- Documentation/technical/api-grep.txt -> grep.h does not have enough
documentation at the moment.
2- Documentation/technical/api-object-access.txt -> sha1-file.c and
object.h have some details
3- Documentation/technical/api-quote.txt -> quote.h has some details.
4- Documentation/technical/api-xdiff-interface.txt -> xdiff-interface.h has
some details.

Signed-off-by: Heba Waly <heba.waly@gmail.com>