trace2: move doc to trace2.h
Move the functions documentation from
Documentation/technical/api-submodule-config.txt to submodule-config.h
as it's easier for the developers to find the usage information beside
the code instead of looking for it in another doc file.

Only the functions documentation section is removed from
Documentation/technical/api-submodule-config.txt as the file is full of
more details that seemed more appropriate to be in a separate doc file
as it is, with a link to the doc file added in the trace2.h.
Also the functions doc is removed to avoid having redundandt info which
will be hard to keep syncronized with the documentation in the header file.

Signed-off-by: Heba Waly <heba.waly@gmail.com>