fix-typo: consecutive-word duplications
Correct unintentional duplication(s) of words, such as "the the",
and "can can" etc.

In certain cases, the repeated words are suspicious but actually ok,
and if some alternative wording exists and might be an improvement,
then the alternative wording is used, for example:
"we didn't do so so far" becomes "we didn't do that so far".

Signed-off-by: ryenus <ryenus@gmail.com>