contrib: change the prompt for am-based rebases
The prompt for am-based rebases was REBASE, while for interactive-based
rebases was REBASE-i.  A while ago, we switched merge-based rebases from
using REBASE-m to REBASE-i via re-implementing the merge backend based
on the interactive backend.  We will soon be changing the default rebase
backend to the interactive one, meaning the default prompt will be
REBASE-i rather than REBASE.  We have also noted in the documentation
that currently am-specific options will be implemented in the
interactive backend, and even the --am flag may eventually imply an
interactive-based rebase.  As such, change the prompt for an am-based
rebase from REBASE to REBASE-a.

Signed-off-by: Elijah Newren <newren@gmail.com>