Documentation/git-log: document accepted line-log diff formats
Currently the line-log functionality (git log -L) only supports
displaying patch output (`-p`, its default behavior) and suppressing it
(`-s`). A check was added in the code to that effect in 5314efaea (line-log:
detect unsupported formats, 2019-03-10) but the documentation was not
updated.

Explicitly mention that `-L` implies `-p`, that patch output can be
suppressed using `-s`, and that all other diff formats are not allowed.

Additionnally, mention that the ':<funcname>' form implies `--function-context`.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>